backend/pyproject.toml
[project]
name = "vlo-backend"
version = "0.1.0"
description = "VLO video editor backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "pydantic",
    "xxhash",
    "python-magic; sys_platform != 'win32'",
    "python-magic-bin; sys_platform == 'win32'",
    "Pillow",
    "httpx",
    "websockets",
    "python-dotenv",
    "numpy",
    "av",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
